Add files via upload
This code first randomly spreads 1000 particles in a 2D plane. Next, it finds out how similar the position of the randomly spread particle is to the actual position of the particle.  
Based on the similarity, it assigns a probability to that particle. Finally, it removes the particles with a lower probability from the set of particles. This algorithm is called as Particle filtering.